+++ packages/driver/src/cypress/script_utils.ts
@@ -43,11 +43,15 @@ const evalScripts = (specWindow, scripts: any = []) => {
   })
 }
 
-const runScriptsFromUrls = (specWindow, scripts) => {
+const runScriptsFromUrls = (specWindow, scripts, projectRoot, specRelativePath, specAbsolutePath) => {
   return Bluebird
   .map<any, any>(scripts, (script) => fetchScript(specWindow, script))
   .map(extractSourceMap)
-  .then((scripts) => evalScripts(specWindow, scripts))
+  .then((scripts) => {
+    $sourceMapUtils.setSourceMapProjectRoot(specRelativePath, specAbsolutePath, projectRoot)
+
+    return evalScripts(specWindow, scripts)
+  })
 }
 
 const appendScripts = (specWindow, scripts) => {
@@ -77,13 +81,18 @@ interface RunScriptsOptions {
   scripts: Script[]
   specWindow: Window
   testingType: Cypress.TestingType
+  projectRoot: string
+  specRelativePath: string
+  specAbsolutePath: string
 }
 
 // Supports either scripts as objects or as async import functions
 export default {
-  runScripts: ({ browser, scripts, specWindow, testingType }: RunScriptsOptions) => {
+  runScripts: ({ browser, scripts, specWindow, testingType, projectRoot, specRelativePath, specAbsolutePath }: RunScriptsOptions) => {
     // if scripts contains at least one promise
     if (scripts.length && typeof scripts[0] === 'function') {
+      $sourceMapUtils.setSourceMapProjectRoot(specRelativePath, specAbsolutePath, projectRoot)
+
       // chain the loading promises
       // NOTE: since in evalScripts, scripts are evaluated in order,
       // we chose to respect this constraint here too.
@@ -94,12 +103,14 @@ export default {
     // in webkit, stack traces for e2e are made pretty much useless if these
     // scripts are eval'd, so we append them as script tags instead
     if (browser.family === 'webkit' && testingType === 'e2e') {
+      $sourceMapUtils.setSourceMapProjectRoot(specRelativePath, specAbsolutePath, projectRoot)
+
       return appendScripts(specWindow, scripts)
     }
 
     // for other browsers, we get the contents of the scripts so that we can
     // extract and utilize the source maps for better errors and code frames.
     // we then eval the script contents to run them
-    return runScriptsFromUrls(specWindow, scripts)
+    return runScriptsFromUrls(specWindow, scripts, projectRoot, specRelativePath, specAbsolutePath)
   },
 }